+++ xyui/avalonia/src/XYUI.Avalonia/Controls/XYUI2/XYUI2-13-Select/XYSelect.cs
@@ -1,0 +1,8 @@
+using Avalonia.Controls;
+
+namespace XYUI.Avalonia.Controls;
+
+public class XYSelect : ComboBox
+{
+    public XYSelect() { Classes.Add("xyui-select"); IsEditable = false; }
+}